+++ crates/barracuda/src/shaders/loss/perceptual_loss_f64.wgsl
@@ -1,4 +1,4 @@
-// Perceptual Loss - Feature-based perceptual loss
+// Perceptual Loss - Feature-based perceptual loss (f64 canonical)
 // Compares high-level features instead of pixels
 // Used in style transfer and super-resolution
 //
@@ -17,14 +17,14 @@ struct Params {
     num_partials: u32, // Number of workgroup partial results
 }
 
-@group(0) @binding(0) var<storage, read> features1: array<f32>;
-@group(0) @binding(1) var<storage, read> features2: array<f32>;
-@group(0) @binding(2) var<storage, read> weights: array<f32>;
-@group(0) @binding(3) var<storage, read_write> loss_buffer: array<f32>;  // partial sums per workgroup
-@group(0) @binding(4) var<storage, read_write> output: array<f32>;
+@group(0) @binding(0) var<storage, read> features1: array<f64>;
+@group(0) @binding(1) var<storage, read> features2: array<f64>;
+@group(0) @binding(2) var<storage, read> weights: array<f64>;
+@group(0) @binding(3) var<storage, read_write> loss_buffer: array<f64>;  // partial sums per workgroup
+@group(0) @binding(4) var<storage, read_write> output: array<f64>;
 @group(0) @binding(5) var<uniform> params: Params;
 
-var<workgroup> shared_loss: array<f32, 256>;
+var<workgroup> shared_loss: array<f64, 256>;
 
 @compute @workgroup_size(256)
 fn main(
@@ -35,7 +35,7 @@ fn main(
     let idx = global_id.x;
     let local_idx = local_id.x;
 
-    var weighted_sq_diff: f32 = 0.0;
+    var weighted_sq_diff: f64 = 0.0;
     if (idx < params.size) {
         let diff = features1[idx] - features2[idx];
         let sq_diff = diff * diff;
@@ -75,7 +75,7 @@ fn compute_mean_loss(
     let local_idx = local_id.x;
     let num_wg = params.num_partials;
 
-    var partial: f32 = 0.0;
+    var partial: f64 = 0.0;
     if (idx < num_wg) {
         partial = loss_buffer[idx];
     }
@@ -93,6 +93,6 @@ fn compute_mean_loss(
 
     if (local_idx == 0u) {
         let total_loss = shared_loss[0];
-        output[0] = total_loss / f32(params.size);
+        output[0] = total_loss / f64(params.size);
     }
 }
