+++ crates/barracuda/src/shaders/math/argmax_f64.wgsl
@@ -1,4 +1,4 @@
-// Argmax - Find indices of maximum values along a dimension
+// argmax_f64.wgsl - Find indices of maximum values along a dimension (f64 canonical)
 //
 // Deep Debt Principles:
 // - Pure WGSL implementation (universal compute)
@@ -12,24 +12,24 @@ struct Params {
     inner_size: u32,
 }
 
-@group(0) @binding(0) var<storage, read> input: array<f32>;
+@group(0) @binding(0) var<storage, read> input: array<f64>;
 @group(0) @binding(1) var<storage, read_write> output: array<u32>;
 @group(0) @binding(2) var<uniform> params: Params;
 
 @compute @workgroup_size(256)
 fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let idx = global_id.x;
-    
+
     if (idx >= params.outer_size * params.inner_size) {
         return;
     }
-    
+
     let outer = idx / params.inner_size;
     let inner = idx % params.inner_size;
-    
+
     var max_value = -1e30; // Very negative number
     var max_idx = 0u;
-    
+
     for (var i = 0u; i < params.dim_size; i++) {
         let input_idx = outer * params.dim_size * params.inner_size + i * params.inner_size + inner;
         let value = input[input_idx];
@@ -38,6 +38,6 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
             max_idx = i;
         }
     }
-    
+
     output[idx] = max_idx;
 }
