+++ crates/barracuda/src/shaders/math/argmax_reduce_f64.wgsl
@@ -1,9 +1,9 @@
-// Argmax Reduction: Find index of maximum value over all elements
+// Argmax Reduction - Find index of maximum value over all elements (f64 canonical)
 // CUDA equivalent: thrust::reduce with argmax operation
 // Algorithm: Tree reduction tracking both value and index
 // Use cases: Global argmax computation
 
-@group(0) @binding(0) var<storage, read> input: array<f32>;
+@group(0) @binding(0) var<storage, read> input: array<f64>;
 @group(0) @binding(1) var<storage, read_write> output: array<u32>;  // Partial results (indices)
 
 struct Params {
@@ -13,7 +13,7 @@ struct Params {
 @group(0) @binding(2) var<uniform> params: Params;
 
 // Shared memory for values and indices
-var<workgroup> shared_values: array<f32, 256>;
+var<workgroup> shared_values: array<f64, 256>;
 var<workgroup> shared_indices: array<u32, 256>;
 
 @compute @workgroup_size(256)
@@ -24,9 +24,9 @@ fn main(
 ) {
     let tid = local_id.x;
     let gid = global_id.x;
-    
+
     // Load data into shared memory
-    var value: f32;
+    var value: f64;
     var index: u32;
     if (gid < params.size) {
         value = input[gid];
@@ -39,7 +39,7 @@ fn main(
     shared_values[tid] = value;
     shared_indices[tid] = index;
     workgroupBarrier();
-    
+
     // Tree reduction in shared memory, tracking indices
     for (var stride = 128u; stride > 0u; stride = stride / 2u) {
         if (tid < stride && (gid + stride) < params.size) {
@@ -47,7 +47,7 @@ fn main(
             let b_val = shared_values[tid + stride];
             let a_idx = shared_indices[tid];
             let b_idx = shared_indices[tid + stride];
-            
+
             if (b_val > a_val) {
                 shared_values[tid] = b_val;
                 shared_indices[tid] = b_idx;
@@ -57,7 +57,7 @@ fn main(
         }
         workgroupBarrier();
     }
-    
+
     // Write partial result (index of max value in this workgroup)
     if (tid == 0u) {
         output[workgroup_id.x] = shared_indices[0];
